Refactor: type 정리
