+++ src/container/questions/components/UpdateQuestionModal/types.ts
@@ -1,4 +1,4 @@
-import { ICategory } from '../../types';
+import { ICategory } from '@/types/categories';
 
 export interface IProps {
   questionId: number;
